Pyhton/britecore/requirements.txt
#cryptography
django~=2.1
djangorestframework
psycopg2
